Add dependency on VPC Policy
It seems as though the Function role is created and the function is
subsequently created before policies that are attached to the role are
created. This means that permissions assigned by certain policies may
not be present when the function is being created. This is improtant for
the creation of ENIs for attaching to subnets.